# speed up parsing for short ipv4s in std::net::Ipv4Addr::from_str