This aims to improve parsing speed for short and empty strings. 

There is a proof of concept based on Criterion here: https://github.com/nkconnor/rust-ipv4-optimization/blob/main/benches/ipv4_fromstr_benchmark.rs

On my machine it offers about a 15X speed-up in the happy case (`len() < 7`). The parser today takes a more similar duration for `"1.1.1"` as it does for `"1.1.1.1"`.

Optimizing for this case is especially helpful for situations where input is a non-null string type. For example, a CSV library may offer non-null string type for an IP Address column. IP Address is a _very common_ element in "big data" and this optimization is likely to be appreciated in such contexts.

Criterion Output:

**Main Parse OK           time:   [21.145 ns 21.181 ns 21.222 ns]**
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe

**Main Parse <7           time:   [15.268 ns 15.286 ns 15.304 ns]**
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

**Patch Parse OK          time:   [21.309 ns 21.506 ns 21.843 ns]**
Found 13 outliers among 100 measurements (13.00%)
  4 (4.00%) high mild
  9 (9.00%) high severe

**Patch Parse <7          time:   [998.09 ps 999.02 ps 1.0001 ns]**