@Mark-Simulacrum 2022-08-09
Can you say more about "the happy case" here? It looks like too short IPs previously went through a longer attempt to parse, but my (maybe wrong) assumption is that such IPs are fairly rare in any data set of IPs.

Are we expecting users to attempt parsing and then fallback on some alternative?

> i'd also like to suggest we move both of these length checks into fn read_ipv4_addr(&mut self) -> Option<Ipv4Addr>. That will short circuit for the IpAddr::from_str case as well.

Presuming this is the same behavior-wise (probably the case, but would need to check call sites), that seems reasonable to me.