@nkconnor 2022-08-09
> Can you say more about "the happy case" here? It looks like too short IPs previously went through a longer attempt to parse, but my (maybe wrong) assumption is that such IPs are fairly rare in any data set of IPs.

That's right, the happy case is where a too short IP is parsed and the parser exits early.

I think IPs with 0 < length < 7 are probably more rare than IPs with length == 0. I was motivated to look at this because we were processing a very large dataset and the question came up whether to do a length check before attempting parsing. Since the parser was slow for the too short IPs, we added the length check. Our dataset had cases of length==0 and 0 < length < 7. 

If we kept the assumption that such IPs are fairly rare, could we say that the patch is still worthwhile? The criterion output shows a small but supposably measurable performance decline for parsing valid IPs.

> Are we expecting users to attempt parsing and then fallback on some alternative?

I would expect this to happen sometimes. In my case, we were provided with a third-party string which may/may not be a valid IP and then writing it in a structured format to a storage archive.

> > i'd also like to suggest we move both of these length checks into fn read_ipv4_addr(&mut self) -> Option. That will short circuit for the IpAddr::from_str case as well.
> 
> Presuming this is the same behavior-wise (probably the case, but would need to check call sites), that seems reasonable to me.

I believe it is. This `read_ipv4_addr` is also used within parsing for IPv6s and SocketAddrs.