@nkconnor 2022-08-09
some tests appear to be failing, looking into it. Also, tested the case of an empty string and the stats are different than above (which tested "1.1.1" .. 0<length<7)
```
current       time:   [4.6269 ns 4.6319 ns 4.6373 ns]
patch         time:   [3.4966 ns 3.4984 ns 3.5004 ns]
```