@nkconnor 2022-08-09
I think tests fail with `SocketAddrs` because the `Parser` state includes the port info, and the max length check (<15) isn't valid when looking at that too. I would have to propagate `AddrKind` down to the parsing function to determine the max length, or leave the max length check where it is.