@nkconnor 2022-08-09
this could/should fail too (ipv6s can be shorter, I think), but it's one example of how we could be consistent in doing length checks across the `net::Parser`. Not sure which direction to go.. I will come back to this.